requirements.txt
asgiref==3.7.2
autocrop==1.3.0
crispy-bootstrap5==2023.10
Django==4.2.7
django-appconf==1.0.6
django-ckeditor==6.7.0
django-crispy-forms==2.1
django-image-cropping==1.7
django-js-asset==2.2.0
easy-thumbnails==2.8.5
numpy==1.26.2
opencv-python-headless==4.8.1.78
Pillow==10.1.0
sqlparse==0.4.4
typing_extensions==4.8.0
